mpy_frozen: refactor module
