+++ tests/checkout_repo/freezer_mpy_test.py
@@ -57,7 +57,8 @@ def test_extract_target_names(path, port, board):
     assert _port == port
 
 
-@pytest.mark.basicgit
+# @pytest.mark.basicgit
+@pytest.mark.slow
 def test_freezer_mpy_manifest(tmp_path: Path, testrepo_micropython: Path, testrepo_micropython_lib: Path):
     "test if we can freeze source using manifest.py files"
     # mpy_path = Path(testrepo_micropython)
@@ -80,7 +81,8 @@ def test_freezer_mpy_manifest(tmp_path: Path, testrepo_micropython: Path, testre
     assert len(scripts) > 10, "expect at least 50 files, only found {}".format(len(scripts))
 
 
-@pytest.mark.basicgit
+# @pytest.mark.basicgit
+@pytest.mark.slow
 def test_freezer_mpy_folders(tmp_path, testrepo_micropython: Path, testrepo_micropython_lib: Path):
     "test if we can freeze source using modules folders"
     mpy_path = testrepo_micropython.as_posix()
